[FIX JENKINS-39842] Open Blue Ocean button should not try to load /activity for a folder (#96)
* Test for JENKINS-39842

Clicking on the "Open Blue Ocean" button while in a classic folder that's not a MBP project folder should bring the user to the main top-level blue ocean page i.e. pipelines

* Another test for JENKINS-39842 for @kzantow

Using folder and job/branch names that are not confusing (special chars etc). BTW I'm leaving the original tests that were using weird chars because I think it's good to have tests that challenge the url encoding etc. Sure that can result in the presence of weird chars, but weird chars are a fact of life, so good to test against them.